feat: Customer BE & FE work flows and UI complete // started mapping out side nav routes

+++ backend/controllers/companyController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require('express-async-handler')
 const bcrypt = require('bcryptjs')
-const { generateToken, userHasPermissions } = require('../utils/auth')
+const { userHasPermissions } = require('../utils/auth')
 
 const Company = require('../models/companyModel')
+const User = require('../models/userModel')
 
 // @desc POST register a new company
 // @route /api/companies
@@ -52,6 +53,7 @@ const registerCompany = asyncHandler(async (req, res) => {
     throw new Error('Invalid company data')
   }
 })
+
 // @desc PUT Edit a company
 // @route /api/companies
 // @access public
@@ -74,6 +76,22 @@ const updateCompany = asyncHandler(async (req, res) => {
   res.status(200).json(updatedCompany)
 })
 
+// @desc GET team by company id
+// @route /api/companies
+// @access public
+const getCompanyTeam = asyncHandler(async (req, res) => {
+  const staff = await User.find({
+    company: req.params.id,
+    isAdmin: false,
+  })
+
+  if (!staff) {
+    res.status(404)
+    throw new Error('Error getting team members, please refresh and try again.')
+  }
+  res.status(200).json(staff)
+})
+
 // @desc GET company by id
 // @route /api/companies
 // @access public
@@ -90,16 +108,23 @@ const getCompany = asyncHandler(async (req, res) => {
     throw new Error('Company not found...')
   }
 })
+
 // @desc DELETE user
 // @route /api/companies
 // @access private
 const deleteCompany = asyncHandler(async (req, res) => {
   const company = await Company.findById(req.params.id)
-  console.log('company :>> ', company)
+  const canDelete = userHasPermissions(req.user.role)
 
   if (!company) {
     res.status(404)
     throw new Error('Company not found')
+  }
+  if (!canDelete) {
+    res.status(404)
+    throw new Error(
+      'You do not have the correct credentials to delete the company.'
+    )
   }
   // delete the company
   await company.deleteOne()
@@ -111,4 +136,10 @@ const deleteCompany = asyncHandler(async (req, res) => {
   })
 })
 
-module.exports = { registerCompany, updateCompany, getCompany, deleteCompany }
+module.exports = {
+  registerCompany,
+  updateCompany,
+  getCompany,
+  deleteCompany,
+  getCompanyTeam,
+}
